more test and mypy fixes
