+++ Tests/AfterTraining/test_after_training.py
@@ -11,8 +11,11 @@ up the most recently run AzureML job from most_recent_run.txt
 """
 
 import os
+import shutil
+import sys
 from pathlib import Path
 
+import numpy as np
 import pytest
 from azureml.core import Model, Run
 
@@ -20,13 +23,17 @@ from InnerEye.Azure.azure_config import AzureConfig
 from InnerEye.Azure.azure_runner import RUN_RECOVERY_FILE
 from InnerEye.Azure.azure_util import MODEL_ID_KEY_NAME, fetch_run, get_comparison_baseline_paths, \
     is_running_on_azure_agent, to_azure_friendly_string
-from InnerEye.Common import fixed_paths, fixed_paths_for_tests
+from InnerEye.Common import common_util, fixed_paths, fixed_paths_for_tests
 from InnerEye.Common.common_util import get_epoch_results_path
 from InnerEye.Common.fixed_paths import DEFAULT_RESULT_IMAGE_NAME
+from InnerEye.Common.fixed_paths_for_tests import full_ml_test_data_path
 from InnerEye.Common.output_directories import OutputFolderForTests
 from InnerEye.ML.common import DATASET_CSV_FILE_NAME, ModelExecutionMode
+from InnerEye.ML.deep_learning_config import CHECKPOINT_FOLDER
+from InnerEye.ML.utils.image_util import get_unit_image_header
 from InnerEye.Scripts import submit_for_inference
-from Tests.ML.util import get_default_workspace
+from Tests.ML.util import assert_nifti_content, get_default_workspace, get_nifti_shape
+from TestsOutsidePackage.test_register_model import SubprocessConfig
 
 
 def get_most_recent_run_id() -> str:
@@ -111,11 +118,7 @@ def test_get_comparison_data(test_output_dirs: OutputFolderForTests) -> None:
     """
     Check that metrics.csv and dataset.csv are created after the second epoch, if running on Azure.
     """
-    most_recent_run = get_most_recent_run_id()
-    azure_config = AzureConfig.from_yaml(fixed_paths.SETTINGS_YAML_FILE,
-                                         project_root=fixed_paths.repository_root_directory())
-    workspace = azure_config.get_workspace()
-    run = fetch_run(workspace, most_recent_run)
+    run = get_most_recent_run()
     blob_path = get_epoch_results_path(ModelExecutionMode.TEST)
     (comparison_dataset_path, comparison_metrics_path) = get_comparison_baseline_paths(test_output_dirs.root_dir,
                                                                                        blob_path, run,
@@ -150,3 +153,64 @@ def test_submit_for_inference(test_output_dirs: OutputFolderForTests) -> None:
     assert not seg_path.exists(), f"Result file {seg_path} should not yet exist"
     submit_for_inference.main(args, project_root=fixed_paths.repository_root_directory())
     assert seg_path.exists(), f"Result file {seg_path} was not created"
+
+
+@pytest.mark.skipif(common_util.is_windows(), reason="Too slow on Windows")
+@pytest.mark.after_training_ensemble_run
+def test_register_and_score_model(test_output_dirs: OutputFolderForTests) -> None:
+    """
+    End-to-end test which ensures the scoring pipeline is functioning as expected when used on a recently created
+    model. This test is run after training an ensemble run in AzureML. It starts "submit_for_inference" via
+    Popen.
+    """
+    azureml_model = get_most_recent_model()
+    assert azureml_model is not None
+    # download the registered model and test that we can run the score pipeline on it
+    model_root = Path(azureml_model.download(str(test_output_dirs.root_dir)))
+    # The model needs to contain score.py at the root, the (merged) environment definition,
+    # and the inference config.
+    expected_files = [
+        *fixed_paths.SCRIPTS_AT_ROOT,
+        fixed_paths.ENVIRONMENT_YAML_FILE_NAME,
+        fixed_paths.MODEL_INFERENCE_JSON_FILE_NAME,
+        "InnerEye/ML/runner.py",
+    ]
+    for expected_file in expected_files:
+        assert (model_root / expected_file).is_file(), f"File {expected_file} missing"
+    checkpoint_folder = model_root / CHECKPOINT_FOLDER
+    assert checkpoint_folder.is_dir()
+    checkpoints = list(checkpoint_folder.rglob("*"))
+    assert len(checkpoints) >= 1, "There must be at least 1 checkpoint"
+
+    # create a dummy datastore to store the image data
+    test_datastore = test_output_dirs.root_dir / "test_datastore"
+    # move test data into the data folder to simulate an actual run
+    train_and_test_data_dir = full_ml_test_data_path("train_and_test_data")
+    img_files = ["id1_channel1.nii.gz", "id1_channel2.nii.gz"]
+    data_root = test_datastore / fixed_paths.DEFAULT_DATA_FOLDER
+    data_root.mkdir(parents=True)
+    for f in img_files:
+        shutil.copy(str(train_and_test_data_dir / f), str(data_root))
+
+    # run score pipeline as a separate process
+    python_executable = sys.executable
+    [return_code1, stdout1] = SubprocessConfig(process=python_executable,
+                                               args=["--version"]).spawn_and_monitor_subprocess()
+    assert return_code1 == 0
+    print(f"Executing Python version {stdout1[0]}")
+    return_code, stdout2 = SubprocessConfig(process=python_executable, args=[
+        str(model_root / fixed_paths.SCORE_SCRIPT),
+        f"--data_folder={str(data_root)}",
+        f"--image_files={img_files[0]},{img_files[1]}",
+        "--use_gpu=False"
+    ]).spawn_and_monitor_subprocess()
+
+    # check that the process completed as expected
+    assert return_code == 0, f"Subprocess failed with return code {return_code}. Stdout: {os.linesep.join(stdout2)}"
+    expected_segmentation_path = Path(model_root) / DEFAULT_RESULT_IMAGE_NAME
+    assert expected_segmentation_path.exists(), f"Result file not found: {expected_segmentation_path}"
+
+    # sanity check the resulting segmentation
+    expected_shape = get_nifti_shape(train_and_test_data_dir / img_files[0])
+    image_header = get_unit_image_header()
+    assert_nifti_content(str(expected_segmentation_path), expected_shape, image_header, [3], np.ubyte)
