+++ Tests/ML/runners/test_compare_against_baselines.py
@@ -42,7 +42,6 @@ def test_perform_score_comparisons() -> None:
     assert list(result.plots.keys()) == [f"{comparison_name}_vs_CURRENT"]
 
 
-# TODO: remove reference to epoch
 @pytest.mark.after_training_single_run
 def test_get_comparison_data(test_output_dirs: OutputFolderForTests) -> None:
     azure_config = get_default_azure_config()
